+++ NEWS.md
@@ -1,9 +1,28 @@
-# semptools 0.2.10.1
+# semptools 0.2.11
+
+## New Features
+
+- Added `set_edge_attribute()` to
+  change any attribute of the edges.
+  (0.2.10.2)
+
+- Added `set_node_attribute()` to
+  change any attribute of the nodes.
+  (0.2.10.3)
+
+- Added `set_edge_color()` to
+  change the colors of edges.
+  (0.2.10.2)
+
+## Miscellaneous
 
 - Revised `change_node_label()` to
   address an issue with `plot.qgraph()`.
   `label.cex` should now be used as
   expected. (0.2.10.1)
+
+- Fixed an R CMD check issue with
+  some links in Rd files. (0.2.10.4)
 
 # semptools 0.2.10